requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
supabase==2.3.0
postgrest==0.16.5
gotrue==2.4.2
realtime==1.0.0
storage3==0.5.3
httpx==0.25.2
python-dateutil==2.8.2
